refactor move stack test to new infra and clean up a little
